+++ packages/react-rerender-test/src/containers/NestedFCPage/NestedFCPage.tsx
@@ -1,18 +1,40 @@
 import { Button } from 'antd';
-import { useContext } from 'react';
+import { useCallback, useContext, useMemo } from 'react';
 
 import { Inspector } from 'src/components/Inspector';
 import { TestPage } from 'src/components/TestPage';
 import { RenderContext, RenderContextUpdater } from 'src/contexts/RenderContext';
 
-import { makeInspectedFC } from '../makeInspectedFC';
+import { makeInspectedFC, usePersistInspectedFC } from '../makeInspectedFC';
 
-const Parent = makeInspectedFC('Parent', () => {
+const ParentNC = makeInspectedFC('Parent', () => {
     useContext(RenderContext);
 
-    const Child = makeInspectedFC('Child', () => <div />);
+    const Child = makeInspectedFC('Nested Child', () => <div />);
 
     return <Child />;
+});
+
+const ParentPNC = makeInspectedFC('Parent', () => {
+    useContext(RenderContext);
+
+    const Child1 = useMemo(() => makeInspectedFC('Memoed Child', () => <div />), []);
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    const Child2 = useCallback(
+        makeInspectedFC('Memoed Child callback', () => <div />),
+        [],
+    );
+
+    const Child3 = usePersistInspectedFC('Persist Child fn', () => <div />);
+
+    return (
+        <>
+            <Child1 />
+            <Child2 />
+            <Child3 />
+        </>
+    );
 });
 
 function Controller(): React.ReactElement {
@@ -28,8 +50,11 @@ function Controller(): React.ReactElement {
 export function NestedFCPage(): React.ReactElement {
     return (
         <TestPage onRenderController={() => <Controller />}>
-            <Inspector>
-                <Parent />
+            <Inspector group="Nested FC">
+                <ParentNC />
+            </Inspector>
+            <Inspector group="Memoed & Persist Nested FC">
+                <ParentPNC />
             </Inspector>
         </TestPage>
     );